- Fix race condition when start watching a mailbox and directly afterwards stop watching it. A BLooper was not suitable to synchronise start and stop watching. Wait till the IDLE command is send before returning the SyncAndStartWatchingMailbox method now. That ensures that a later StopWatchingMailbox call find the maibox in an expected watching state.
There is one thread (BLooper) to handle new commands and one watcher thread which is just listening at the server port for updates. The race condition occurred for example when a sync/watching and a fetch body message are send to the looper. The sync message just triggered the IDLE command in the watcher thread. In the meantime the fetch body command send a DONE command, because the IDLE command has not be send at this time the watcher keeps watching. 

- fix int32 -> ssize_t thanks Axel and Stippi
- clean up



git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@40919 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/preferences/mail/DNSQuery.h
@@ -42,7 +42,7 @@ public:
 
 private:
 		void			_Init(const void* buf, size_t size);
-		int32			_ReadStringAt(BString& string, off_t pos);
+		ssize_t			_ReadStringAt(BString& string, off_t pos);
 
 		off_t 			fWritePosition;
 		off_t 			fReadPosition;
